requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
attrs==25.3.0
bitarray==3.6.0
blinker==1.9.0
certifi==2025.8.3
charset-normalizer==3.4.2
ckzg==2.1.1
click==8.2.1
colorama==0.4.6
cytoolz==1.0.1
eth-account==0.13.7
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.7.0
eth-rlp==2.2.0
eth-typing==5.2.1
eth-utils==5.3.0
eth_abi==5.2.0
Flask==3.1.1
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
frozenlist==1.7.0
greenlet==3.2.3
hexbytes==1.3.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
multidict==6.6.3
packaging==25.0
parsimonious==0.10.0
propcache==0.3.2
py-solc-x==2.0.4
pycryptodome==3.23.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMySQL==1.1.1
pyunormalize==16.0.0
pywin32==311
regex==2025.7.34
requests==2.32.4
rlp==4.1.0
SQLAlchemy==2.0.42
toolz==1.0.0
types-requests==2.32.4.20250611
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
web3==7.12.1
websockets==15.0.1
Werkzeug==3.1.3
yarl==1.20.1
flask
flask_sqlalchemy
flask_cors
pymysql
